# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: AI Review by Digital Hammerr API — maintained V1 contract
  version: 1.0.0-impl.1
  summary: The endpoints that actually exist, documented from the handlers.
  description: |
    This is the **maintained** contract. `docs/spec/08_OpenAPI_v1.yaml` is the frozen delivered
    artefact and remains the contract of record for V1 *scope* — every path it lists is still in
    scope. It is not usable as a validation contract: OPEN-04 in `docs/SPEC_AMENDMENTS.md` records
    that most of its operations declare a bare `200` with no schema, `/auth/login` has no request
    body at all, and only `SignupRequest`, `ReviewGeneration` and `Error` are defined. The pack's
    own required "Contract: OpenAPI response validation" suite cannot be written against it.

    Everything here was read out of `apps/web/app/api/v1/**/route.ts` rather than copied from the
    skeleton, and every operation carries `x-implemented-by` naming the handler it describes. Field
    constraints come from the Zod contracts in `packages/contracts/src`, which are authoritative:
    where this YAML would disagree with them, this YAML is wrong.

    Two things to know before reading:

    - **`x-error-codes`** on each error response lists the `error.code` values that handler can
      actually emit at that status. It is an extension rather than a narrowed schema because the
      envelope is a fixed shape and repeating a four-level-deep `enum` override on some sixty
...

...
    the delta.
servers:
...
security:
  - sessionCookie: []
...
paths:
  /auth/signup:
    post:
      operationId: signup
      tags: [Auth]
      summary: Create an owner account plus its DRAFT tenant
      description: |
        AUTH-01. Writes the user, a DRAFT shell business and a FREE subscription in one
        transaction, so an invalid request cannot leave a partial active tenant behind (AC-001),
        then issues a session cookie and reports where to land.

        `422` covers every rejection: malformed JSON, a Zod failure, a password that fails the
        strength rules in `packages/core`, an unnormalizable mobile, and an email already in use.
        The duplicate-email case is a deliberate disclosure — AUTH-01 specifies that state and a
        signup form cannot work without it. The neutral-response rule belongs to
        `/auth/forgot-password`, not here.

        **Not rate limited today.** No limiter call exists in this handler, unlike login and
        forgot-password. Recorded in `docs/openapi/README.md` under known gaps.
      security: []
      x-implemented-by: apps/web/app/api/v1/auth/signup/route.ts
      parameters:
        - $ref: '#/components/parameters/OriginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: Account, DRAFT tenant and session created.
          headers:
            Set-Cookie: { $ref: '#/components/headers/SessionCookieSet' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignupResponse' }
        '403':
          $ref: '#/components/responses/CsrfRejected'
        '422':
          description: |
            Malformed body, field validation, weak password, invalid mobile, or email already
            registered. `details.fields` names the offending fields only — never their values
            (AUTH-01-03, AC-030).
          x-error-codes: [VALIDATION_FAILED]
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                emailTaken:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      message: An account with that email already exists.
                      request_id: 6b1f2c40-9a3e-4d21-8f0b-1c7e5a9d4e11
                      details: { fields: [email] }
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dh_session
      description: |
        Opaque session token. `SESSION_COOKIE_NAME` in `packages/config/src/env.ts` defaults to
        `dh_session`; a deployment may rename it. Set `HttpOnly`, `SameSite=Lax`, `Path=/`, and
        `Secure` in production (`13_Security_Privacy_Compliance.md`: no auth token in
        localStorage). Only the token's SHA-256 is stored server-side, in `sessions`
        (AMENDMENT-001) — the raw value exists only in the cookie.

        Authenticated handlers resolve the tenant from this cookie and compare resource ownership;
        no endpoint accepts a business id (RBAC rule 2, AC-003). An expired, revoked and absent
        cookie are indistinguishable in every response, so a caller cannot probe whether a token
        was ever valid.
  parameters:
    OriginHeader:
      name: Origin
      in: header
      required: true
      description: |
        CSRF defence for cookie-authenticated mutations (`apps/web/lib/csrf.ts`). Every browser
        sends `Origin` on a cross-origin mutation, so its **absence** on a state-changing request
        is itself rejected, and a value that is not `APP_BASE_URL` (plus localhost outside
        production) is rejected too. Custom tenant domains are deliberately not accepted: they
        serve the public renderer only, and widening this would make any hostname a tenant can
        point at us a trusted origin for the dashboard.

        Public endpoints are exempt — they carry no session and no authority — as is
        `DELETE /auth/login`, which is a gap rather than a decision.
      schema: { type: string, format: uri }
      example: https://review.digitalhammerr.com
...
  headers:
...
    SessionCookieSet:
      description: |
        Sets the session cookie: `HttpOnly`, `SameSite=Lax`, `Path=/`, `Secure` in production, and
        `Expires` from the session's own expiry (longer when `remember_me` was sent).
      schema: { type: string }
...
  responses:
...
    CsrfRejected:
      description: |
        The `Origin` header was missing or did not match this application. Only reachable on
        mutating methods — `requireTenant` runs the check before it looks for a session, so a
        forged cross-site request is refused without a database read.
      x-error-codes: [FORBIDDEN]
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: FORBIDDEN
              message: Request rejected.
              request_id: 8a41d0e2-1c93-4f57-b8d6-3e0a2c5f9b17
...
    InternalError:
      description: Unexpected failure. Safe message only; never a provider error or stack trace.
      x-error-codes: [INTERNAL_ERROR]
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: |
        The single error envelope from `docs/spec/23_API_Error_Codes.md`, produced by
        `apps/web/lib/api-error.ts`. `request_id` is a fresh UUID on every error so a
        customer-facing message can be traced to a log line without the message itself carrying
        anything. Provider errors, stack traces and internal identifiers never appear in it
        (AC-030).
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, request_id]
          additionalProperties: false
          properties:
            code:
              type: string
              description: |
                One of the codes in `23_API_Error_Codes.md`. Each error response in this document
                lists the subset its handler emits under `x-error-codes`.
              enum:
                - AUTH_REQUIRED
                - AUTH_INVALID_CREDENTIALS
                - AUTH_RATE_LIMITED
                - FORBIDDEN
                - TENANT_SCOPE_VIOLATION
                - VALIDATION_FAILED
                - RESOURCE_NOT_FOUND
                - SLUG_UNAVAILABLE
                - BUSINESS_NOT_ACTIVE
                - REVIEW_DESTINATION_INVALID
                - PLAN_QUOTA_EXHAUSTED
                - FAIR_USE_THROTTLED
                - PUBLIC_RATE_LIMITED
                - AI_PROVIDER_UNAVAILABLE
                - AI_OUTPUT_REJECTED
                - QR_NOT_FOUND
                - QR_DISABLED
                - DOMAIN_ALREADY_CLAIMED
                - DOMAIN_VERIFICATION_PENDING
                - PAYMENT_VERIFICATION_FAILED
                - PAYMENTS_NOT_CONFIGURED
                - SUBSCRIPTION_NOT_ACTIVE
                - UPLOAD_INVALID
                - ADMIN_REASON_REQUIRED
                - INTERNAL_ERROR
            message:
              type: string
              description: Safe, user-facing, already written for display. Never a raw error.
            request_id: { type: string, format: uuid }
            details:
              type: object
              additionalProperties: true
              description: |
                Present only where a client can act on it. In practice: `fields` (an array of
                field names, or a name-to-message map on public feedback) and `missing` (publish
                prerequisites).
    SignupRequest:
      type: object
      description: |
        `signupRequest` in `packages/contracts/src/auth.ts`. `accept_terms` is a literal `true`,
        not a boolean: an unchecked box is not a valid signup (AUTH-01).
      required: [full_name, email, mobile, password, accept_terms]
      properties:
        full_name: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string, format: email }
        mobile:
          type: string
          minLength: 8
          maxLength: 20
          description: |
            Normalised to E.164 by `normalizePhone`; D-002 expects a bare 10-digit Indian mobile.
            A value that will not normalise is a `422` on `mobile`.
        password:
          type: string
          minLength: 12
          maxLength: 256
          format: password
          description: |
            Length is checked by Zod; `validatePasswordStrength` in `packages/core` additionally
            rejects common and low-variety passwords. Never logged and never echoed (AUTH-01-03).
        accept_terms: { type: boolean, const: true }
    SignupResponse:
      type: object
      required: [next, onboarding_required]
      properties:
        next:
          type: string
          description: Where to land. `/app` for an owner; `landingPathFor` decides.
          example: /app
        onboarding_required:
          type: boolean
          const: true
          description: |
            Always true — signup creates a DRAFT shell tenant, so the wizard always follows
            (AUTH-01-01, ONB-01).
...